abstract out createListControl
